adjust type follow wm type for vd51 to 54

+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd53Dto.java
@@ -2,11 +2,13 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Vd53Dto extends Vd5Base {
-    private double nilaiDitambahkan;
+    private BigDecimal nilaiDitambahkan;
 }
